reworked slightly how JobTransfers compute CRC
we were computing the crc on the source and destination side and then comparing the destination side to the local cached checksum in the master's 
filelist.  The fatal flaw here is that the local cached version could be out of sync with what is on the source's filesystem, essentially creating an 
infinite loop.  To solve this, I trust that the source crc value computed on the transfer is correct (it is, it's reading directly from the 
drive) and compare that to the destination crc