Revert [2045] and try a new approach.
Since 3.0 does not include checksums in an initial merge and instead relies on lazy instantiation of these in the VFS and retrieves them from the slave on first demand the logic will defer to the slave when the checksum is 0. This can cause a problem as files are created in the VFS for an upload before the listen/connect is issued to the slave (and a physical file consequently created on the slave), at this point the pending file in the VFS has a 0 checksum, if another thread tries to get a checksum for this file (due to say an MLST in the parent directory) it will issue a request to the slave for it (due to the lazy initialisation) and cause the slave to drop when it cannot find the requested file.

This behaviour has been changed to only retrieve the checksum from the slave when it is 0 in the VFS and the size of the file in the VFS is non-zero, this should prevent the above issue hopefully.

This is to address #266