A general cleanup of the handling of the abort command, failed transfers, timed out control connections during transfer and similar. These changes aim to keep the slave physical filesystem and master VFS in step under all these conditions, a new config option has been introduced in master.conf to control whether files should be deleted on abort (but whichever way this is set it should be consistant now).
Some changes to zipscripts to prevent them trying to parse files which are 0 byte and only exist in the master VFS.

Some changes to stats recording to fix a free credits bug and ensure credits are always updated appropriately for aborted/truncated transfers.

This also fixes the leak of transfer pointers mentioned in #235.