New LinkManager Plugin:
This is to replace the current setup using org.drftpd.commands.zipscript.links*

It would be wise to disable the old plugin(s) before enabling this one, or by changing the following in zipscript.conf to:

incomplete.links=false
incomplete.zip.links=false

I have used the command "site fixlinks2" for now to create missing links, as "site fixlinks" is currently used for the other setup.

Technical notes:  It uses the VFS events for rename/delete which insures the destination exists.  I have not read anything else from the source INode so FileNotFound will not occur.