Fix: keep sending pings until timeout, not only 1 ping
Before this commit we did only send 1 ping (aka pingIndex == null).
If for example the first ping gets lost, the slave will timeout if no other commands are executed.
This because pingIndex will never be 'null' again, so it will never do a new ping.
We fixed this here by running ping until we reach timeout