Fix: possible leak in slave index pool
If a ping was sent to slave but response never received back(10s timeout on read) a new ping would be sent. But index used for lost ping was never returned to the slaves index pool. If this was to happen 256 times index pool would be exhausted and remoteslave object stuck in an infinite loop on master.