Initial commit to replace XMLEncoder/XMLDecoder with json-io
jdk8 and newer has shown issues with XMLEncoder/XMLDecoder and how its used in Drftpd. This is an attempt to bypass this problem. The resulting json file will in most cases use less disk space than the xml equivalent and in some cases as much as half. The files are small but in large number it could make a big difference. All uses of XMLEncoder/XMLDecoder are replaced with json-io except in builder.

* Requirements
  * Use jdk7 to convert old xml files to json before migrating to jdk8+.
  * Remove files/ and index/ directories to start with a fresh VFS.

* User and slave files, as well as nukelog.xml will be converted from xml to json automatically.
* If plugin metadata is saved to VFS and then reloaded/removed, currently loading the VFS inode will fail and a new empty inode will be created. Please remerge or recreate VFS in this case.

Consider this a beta, not recommended for production use. Take backup of users, slaves and nukelog.xml to be able to restore in case of failure.