New SiteBot announce plugin
Removed nfo announce from SFVAnnouncer.
Cleaned up theme properties files in zipscript plugins.
Made some changes to the AbstractAnnouncer and other SiteBot classes to be able to access SiteBot config path needed in new plugin.

New announcer added org.drftpd.plugins.sitebot.announce.ext
This announcer is added to be able to announce the upload of any file type and not just nfo.
File types are specified in irc.conf and theme needs to be updated to included lines for all extensions that is going to be announced.
If changes are made to the list of extensions in irc.conf, a plugin restart of ext announcer(or master restart) is needed.