CRC on remerge and synchronization changes
Added option(default true) to CRC files on remerge before sending to master to be able to verify changes on slave side other than just file size.
If adding files outside of drftpd this also helps in first entry load time for zipscript directories as master then know the CRC for the file and doesn't have to query slave.
It will however add remerge time on slave. Depending on number of slaves remerging at the same time, master load etc this might or might not add to the total remerge time on master.

This commit also includes some synchronization changes on slave side.
A sending transfer(download) could get stuck in OutputStream.write() for unknown reason and its not possible to set a timeout on write. If this happens all downloads after this also get stuck as on download it will go through all active transfers to see if there are any matching uploads for the same file. This to be able to download a currently uploading file. Download then gets stuck waiting to get access to the other stuck download until it terminates.
Change removes a couple of synchronization blocks and maps/sets are replaced with concurrent counterparts for transfer list and rename queue.
The change is tested to the best of my ability but if you notice strange behaviors linked to this create an issue with a reference to this commit.

If log mode is set to info it now logs negotiated SSL protocol and cipher for the transfer.