Remerge CRC
Changed behavior of crc on remerge added in 97ac8d0c4e291c0ac75e9daaddd3d1a20142c474
Instead of crc checking all files on slave on every remerge(very cpu/disk intensive) the control is moved to master side. Master checks if crc is missing i VFS and if so adds it to a crc queue. A separate crc thread is started to process this queue and queries slave for checksum for every file processed. When both remerge and crc queue is drained the remerge is considered finished.