Slave shutdown and queue sizes
This commit adds checks to slaves remerge thread(s) to see if it should gracefully quit if slave has been thrown offline or has been shut down for some other reason.
Solves #21

An update to site slave and irc remergequeue commands is added to list the size of the different slave queues, rename, remerge and remerge-crc.