JSON-IO issue with non ASCII chars
JSON-IO was unable to correctly write non ASCII characters to disk. JSON-IO already wraps InputStream/OutputStream in buffered readers/writers and so when dr also wrapped the streams the encoding got broken.
Removed the extra wrapping of streams in dr and used the "new" java try with resource to automatically close the streams/readers/writers.